Move all XAI report code to report.xai
